`@remotion/renderer`: Factor out streaming into separate file
